code re-use start-app
